moves some basic utilities to the base Covariate module
